Rename filesystem_wrapper to filesystem_wrappers.
