+++ src/data_generators/data_design.py
@@ -23,10 +23,10 @@ from os import listdir
 from os.path import exists, isdir, join
 from typing import List
 
-from wfa_planning_evaluation_framework.filesystem_wrapper import (
+from wfa_planning_evaluation_framework.filesystem_wrappers import (
     filesystem_wrapper_base,
 )
-from wfa_planning_evaluation_framework.filesystem_wrapper import (
+from wfa_planning_evaluation_framework.filesystem_wrappers import (
     filesystem_pathlib_wrapper,
 )
 from wfa_planning_evaluation_framework.data_generators.data_set import DataSet
