+++ src/data_generators/tests/data_design_test.py
@@ -18,7 +18,7 @@ from unittest.mock import patch
 from tempfile import TemporaryDirectory
 
 import cloudpathlib.local
-from wfa_planning_evaluation_framework.filesystem_wrapper import (
+from wfa_planning_evaluation_framework.filesystem_wrappers import (
     filesystem_cloudpath_wrapper,
 )
