+++ src/data_generators/tests/data_set_test.py
@@ -19,7 +19,7 @@ from tempfile import TemporaryDirectory
 from unittest.mock import patch
 
 import cloudpathlib.local
-from wfa_planning_evaluation_framework.filesystem_wrapper import (
+from wfa_planning_evaluation_framework.filesystem_wrappers import (
     filesystem_cloudpath_wrapper,
 )
