+++ src/filesystem_wrappers/tests/filesystem_pathlib_wrapper_test.py
@@ -18,7 +18,7 @@ import os
 import glob
 from absl.testing import absltest
 
-from wfa_planning_evaluation_framework.filesystem_wrapper import (
+from wfa_planning_evaluation_framework.filesystem_wrappers import (
     filesystem_pathlib_wrapper,
 )
